pacemaker: created openstack_pacemaker_primitive
This change is meant to provide a default value for pacemaker ops. In
particular we are interested in providing a default for 'op on-fail'
actions.

With Pacemaker we are not able to configure such behavior. The approach
here is to provide a wrapper definition around pacemaker_primitive that
can read that default value from the Pacemaker barclamp. For that reason
this change depends on a pull request in the crowbar-ha repository
addressing this.

(cherry picked from commit 859e2d6d7f540064f974bc0a5879bdee6f54c991)